requirements.txt
requests==2.27.1
python-dotenv==0.19.2
python_telegram_bot==13.11